Extract test call
